Update <22.03.2025>
On the plans for the post rating system:
A more technical foundation/representation must be laid for the DB relations. As of now, there is another entity planned for the ratings, PostRatings, essentially a representation for an M:M table between the *users* and *posts* tables with a Tinyint slapped on top to indicate the individual user ratings for each Post entry.
As for the Post objects' *points* field, it'll likely be controlled on a per-click basis from the Service layer, as in, clicking a positive or negative rating, will both create a record in the new *post_ratings* table AND increment/decrement the points.
Note: *post_ratings* table can have its Tinyint column changed to boolean. If we like a post, we create a record inside with this boolean set to true, if we dislike - false, if we remove our reaction, we'll delete the record. Should be more memory and performance efficient. This necessitates frequent use of a hypothetical repo method to return nulls, need to check if that's viable(probably is).
Note: Seems like a good job for a scheduler to verify.

Changes with notes:

Models finished*:
User, Thread, Post^

Repos finished**:
UserRepo, ThreadRepo, PostRepo

Services started:
UserService and ThreadService

^The posts table, and therefore model, could be altered to include a boolean is_hidden, for admins to exert control over the forums. Still on debate but likely will go in.

*relations and Collection-based fields need to be rechecked and completed. But the majority of the proofing is finished.

**Repositories are in a working state but unknown uses for now. It's likely more queries will have to be implemented.

General note: Install the lombok plugin on the PC! :)