Update <26.03.2025>
Work begins using the front-end.

! Status of models still not rigorously tested.
* Repositories have been expanded and work for now.

Changes with notes:

spring-security has been configured to work well in terms of accessing pages. The actual log in process is yet to be tested.
Small note here, for debugging purposes, it'd probably be a good idea to put something like "/**" in the .requestMatchers().

HTML with Thymeleaf forms added. Basic testing has been done with /register, seems to add entries in the DB properly. /login is still on the table for testing whether it logs you in or not, i.e. is there a difference if the login is successful.

Work on controllers began. For now at least, all @Controller annotated classes seem to work fine. Problem: While controllers seem to work fine, REST-using HTML pages need more advanced processing. This seems like a good use case for fragments where we render and need to load multiple thread objects, such as the main page for example.
Note: Advanced processing on the front-end will need JS. Probably a good idea to start classifying things about now.

Final thoughts: Proper detection and function when a user is logged in on a device needs to be studied. We don't want sending random connections user data, this should be spring-security's job.
IDEA: User-only pages, selected with .authenticated, require users per definition. Or rather, user DATA in specific fields. So we could use this data with repo methods such as findByUsername[supposing every page displays the username, which we will], and then find the specific user data. This should work but seems very unsafe, in fact, terrible. Possibly without JWT for the moment. Likely a session-based security chain. This does mean that WebMvcConfiguration.java might need to be further improved.