Micro-service progress
As it seems, micro-service related functionality, both in it and in the monolith, is finished. Only the AvatarService methods need to be more error-proof.
What's missing is automated testing.

Note: .css for the fragment is changed and made more appealing. Still bad, but it was quick to make anyway.
Note: AvatarController.java is an absolutely useless class, likely without any future use.